Refactor code

+++ backend/controllers/product.controller.js
@@ -51,6 +51,10 @@ export const getProducts = async (req,res) => {
 
  export const deleteProduct = async (req,res) => {
     const {id} = req.params
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message: "Invalid Product Id"});
+    }
     
     try{
         await Product.findByIdAndDelete(id);
@@ -59,7 +63,7 @@ export const getProducts = async (req,res) => {
     }catch(error){
         console.error("Error in Delete product:", error.message);
 
-        res.status(404).json({ success: false, message: "Product not found"});
+        res.status(500).json({ success: false, message: "Server Error"});
 
     }
